pyproject.toml

[tool.poetry]
name = "sustainability-extension"
version = "0.0.1"
description = ""
authors = ["Eugene Kang, Hugo Lawes, Ayumi Madaschi"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.7"
beautifulsoup4 = "^4.11.2"
requests = "^2.28.2"
djangorestframework = "^3.14.0"
django-elasticsearch-dsl = "^7.3"
django-extensions = "^3.2.1"
Werkzeug = "^2.2.3"
pyOpenSSL = "^23.0.0"
django-elasticsearch-dsl-drf = "^0.22.5"
django-cors-headers = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
